refactor: decompose commands.ts into 5 focused modules
Extract cohesive groups of functions from the 2,247-line commands.ts
god-file into focused modules:

- commands-prefs-wizard.ts (~600 lines): TUI preferences wizard,
  all configure* functions, serialization helpers
- commands-config.ts (~80 lines): TOOL_KEYS, loadToolApiKeys,
  API key management
- commands-inspect.ts (~80 lines): InspectData type,
  formatInspectOutput, DB diagnostics
- commands-maintenance.ts (~200 lines): cleanup branches/snapshots,
  skip, dry-run handlers
- commands-handlers.ts (~335 lines): doctor, steer, capture, triage,
  knowledge, run-hook, update, skill-health handlers

commands.ts retains registerGSDCommand (router), dispatchDoctorHeal,
fireStatusViaCommand, projectRoot, handleStatus, handleVisualize,
handleSetup, showHelp, and re-exports from all sub-modules to
preserve the public API surface.
